Update app.yaml to use connection name (#679)
We've moved away from concatenating project:region:instance and towards using the instance connection name, which is now available not only from the console but also from gcloud. I'm in the process of adding the text to go along with this change, but already the proxy invocation uses this so we're no worse off if we publish this change.

Thanks!