[jobs] testing: reduce test flakiness (#3278)
* [jobs] testing: reduce test flakiness

fixes #2945
fixes #2818

There were 8 tests which has `time.sleep(10)` for waiting data
propagation or whatever. I chose to use `@eventually_consistent.call`
instead of sleep in those tests.

Bonus point is now the tests are significantly faster because it doesn't
always wait for 10 seconds.

The downside is it will take a long time on failure. We may want to
specify the retry count on the eventualy consistent decorator once the
following issue is fixed:

https://github.com/GoogleCloudPlatform/python-repo-tools/issues/25

* style fix